Refactor CLI structure and update entry points
Summary:
- Moved the main application logic from `run.py` to a new `cli.py` file for better organization.
- Updated `pyproject.toml` to reflect the new entry point for the CLI.
- Adjusted `__main__.py` to import the main function from `cli.py`.
- This change enhances the modularity of the codebase and prepares for future CLI enhancements.

Files modified: `pyproject.toml`, `run.py`, `tritonparse/__main__.py`, `tritonparse/cli.py` (new file).
