+++ tritonparse/__main__.py
@@ -1,4 +1,4 @@
-from .run import main
+from .cli import main
 
 
 if __name__ == "__main__":
